[REFACTOR,CONFIG] update preference presentation in setting: updated the way that the preferences are propogated to the settings menu and also improved the concurency of the login

+++ src/main/kotlin/me/code4me/services/modules/telemetry/TypingSpeed.kt
@@ -32,8 +32,8 @@ class TypingSpeed() : PluginModule {
         val cpsKey = Record.key<Double>("typing_speed_cps")
         val windowSize =
             PrefState
-                .getPreferenceValue(getPreferenceId(), "telemetry.typing_speed.window_size")?.
-                toInt() ?: 10
+                .getPreferenceValue(getPreferenceId(), "telemetry.typing_speed.window_size")
+                ?.toInt() ?: 10
         val cps = trackingService.getTypingSpeed(windowSize).toDouble()
 
         record.put(cpsKey, cps)
